Fix title adornment & namespace, improve snapshots and fix contrast
