+++ app/plugins/adp-backend/package.json
@@ -38,7 +38,6 @@
     "knex": "^3.1.0",
     "node-fetch": "^2.6.7",
     "npm": "^10.3.0",
-    "url-slug": "^4.0.1",
     "winston": "^3.2.1",
     "yn": "^4.0.0",
     "zod": "^3.23.5"
