+++ app/plugins/adp-backend/src/utils/index.test.ts
@@ -1,30 +1,6 @@
 import type { DeliveryProject } from '@internal/plugin-adp-common';
-import {
-  createName,
-  getCurrentUsername,
-  createGithubTeamDetails,
-} from './index';
+import { getCurrentUsername, createGithubTeamDetails } from './index';
 import express from 'express';
-
-describe('createName', () => {
-  it('replaces spaces with dashes and converts to lowercase', () => {
-    const input = 'Example Name';
-    const expected = 'example-name';
-    expect(createName(input)).toBe(expected);
-  });
-
-  it('handles strings longer than 64 characters', () => {
-    const longName = 'a'.repeat(70);
-    const expected = 'a'.repeat(64);
-    expect(createName(longName)).toBe(expected);
-  });
-
-  it('removes extra spaces within the name', () => {
-    const input = 'Example   Name With  Spaces';
-    const expected = 'example-name-with-spaces';
-    expect(createName(input)).toBe(expected);
-  });
-});
 
 describe('getCurrentUsername', () => {
   const mockIdentityApi = {
