+++ app/plugins/adp-common/src/index.ts
@@ -13,3 +13,4 @@ export * from './types/validation';
 export * from './types/arms-length-body';
 export * from './types/delivery-project';
 export * from './types/delivery-programme';
+export * from './util/createName';
