+++ app/plugins/adp-common/src/util/createName.test.ts
@@ -1,0 +1,21 @@
+import { createName } from './createName';
+
+describe('createName', () => {
+  it('replaces spaces with dashes and converts to lowercase', () => {
+    const input = 'Example Name';
+    const expected = 'example-name';
+    expect(createName(input)).toBe(expected);
+  });
+
+  it('handles strings longer than 64 characters', () => {
+    const longName = 'a'.repeat(70);
+    const expected = 'a'.repeat(64);
+    expect(createName(longName)).toBe(expected);
+  });
+
+  it('removes extra spaces within the name', () => {
+    const input = 'Example   Name With  Spaces';
+    const expected = 'example-name-with-spaces';
+    expect(createName(input)).toBe(expected);
+  });
+});
