+++ app/plugins/adp/src/components/ALB/CreateAlbButton.test.tsx
@@ -1,27 +1,25 @@
-import type { AlertApi} from '@backstage/core-plugin-api';
+import type { AlertApi } from '@backstage/core-plugin-api';
 import { alertApiRef } from '@backstage/core-plugin-api';
 import React from 'react';
-import type { ArmsLengthBodyApi} from './api';
+import type { ArmsLengthBodyApi } from './api';
 import { armsLengthBodyApiRef } from './api';
 import { render as testRender, waitFor } from '@testing-library/react';
 import { TestApiProvider } from '@backstage/test-utils';
 import type { CreateAlbButtonProps } from './CreateAlbButton';
 import { CreateAlbButton } from './CreateAlbButton';
 import userEvent from '@testing-library/user-event';
-import type { AlbFields} from './AlbFormFields';
+import type { AlbFields } from './AlbFormFields';
 import { AlbFormFields, emptyForm } from './AlbFormFields';
 import { act } from 'react-dom/test-utils';
 import type { ValidationError as IValidationError } from '@internal/plugin-adp-common';
-import { ValidationError } from '../../utils';
+import { SnapshotFriendlyStylesProvider, ValidationError } from '../../utils';
 import type * as PluginPermissionReactModule from '@backstage/plugin-permission-react';
 import type * as DialogFormModule from '../../utils/DialogForm';
 
 const usePermission: jest.MockedFn<
   typeof PluginPermissionReactModule.usePermission
 > = jest.fn();
-const DialogForm: jest.MockedFn<
-  typeof DialogFormModule.DialogForm
-> = jest.fn();
+const DialogForm: jest.MockedFn<typeof DialogFormModule.DialogForm> = jest.fn();
 
 const fields: AlbFields = {
   alias: 'abc',
@@ -256,7 +254,9 @@ function setup() {
             [armsLengthBodyApiRef, mockArmsLengthBodyApi],
           ]}
         >
-          <CreateAlbButton {...props} />
+          <SnapshotFriendlyStylesProvider>
+            <CreateAlbButton {...props} />
+          </SnapshotFriendlyStylesProvider>
         </TestApiProvider>,
       );
       await waitFor(() => expect(result.baseElement).not.toBeEmptyDOMElement());
