+++ app/plugins/adp/src/components/DeliveryProgramme/DeliveryProgrammeFormFields.test.tsx
@@ -1,33 +1,27 @@
 import React from 'react';
-import type {
-  DeliveryProgrammeFields} from './DeliveryProgrammeFormFields';
+import type { DeliveryProgrammeFields } from './DeliveryProgrammeFormFields';
 import {
   DeliveryProgrammeFormFields,
   emptyForm,
 } from './DeliveryProgrammeFormFields';
-import type {
-  RenderResult} from '@testing-library/react';
+import type { RenderResult } from '@testing-library/react';
 import {
   fireEvent,
   render as testRender,
   waitFor,
 } from '@testing-library/react';
-import type {
-  FieldPath,
-  FieldValues,
-  UseFormReturn} from 'react-hook-form';
-import {
-  useForm,
-} from 'react-hook-form';
+import type { FieldPath, FieldValues, UseFormReturn } from 'react-hook-form';
+import { useForm } from 'react-hook-form';
 import { act } from 'react-dom/test-utils';
-import type { ArmsLengthBodyApi} from '../ALB/api';
+import type { ArmsLengthBodyApi } from '../ALB/api';
 import { armsLengthBodyApiRef } from '../ALB/api';
 import { TestApiProvider } from '@backstage/test-utils';
-import type { ErrorApi} from '@backstage/core-plugin-api';
+import type { ErrorApi } from '@backstage/core-plugin-api';
 import { errorApiRef } from '@backstage/core-plugin-api';
 import userEvent from '@testing-library/user-event';
+import { SnapshotFriendlyStylesProvider } from '../../utils';
 
-describe('AlbFormFields', () => {
+describe('DeliveryProgrammeFormFields', () => {
   it('Should render all fields correctly', async () => {
     const { mockAlbApi, render } = setup();
 
@@ -193,7 +187,9 @@ function setup() {
             [errorApiRef, mockErrorApi],
           ]}
         >
-          <Sut context={context} defaultValues={defaultValues} />
+          <SnapshotFriendlyStylesProvider>
+            <Sut context={context} defaultValues={defaultValues} />
+          </SnapshotFriendlyStylesProvider>
         </TestApiProvider>,
       );
       await waitFor(() => expect(result.baseElement).not.toBeEmptyDOMElement());
