+++ app/plugins/adp/src/components/DeliveryProgramme/DeliveryProgrammeViewPageComponent.test.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import { Button } from '@material-ui/core';
 import { TestApiProvider, renderInTestApp } from '@backstage/test-utils';
-import type { DeliveryProgrammeApi} from './api';
+import type { DeliveryProgrammeApi } from './api';
 import { deliveryProgrammeApiRef } from './api';
-import type { ErrorApi} from '@backstage/core-plugin-api';
+import type { ErrorApi } from '@backstage/core-plugin-api';
 import { errorApiRef } from '@backstage/core-plugin-api';
 import { DeliveryProgrammeViewPageComponent } from './DeliveryProgrammeViewPageComponent';
 import { waitFor } from '@testing-library/react';
@@ -12,9 +12,14 @@ import type { DeliveryProgramme } from '@internal/plugin-adp-common';
 import { entityRouteRef } from '@backstage/plugin-catalog-react';
 import type * as EditDeliveryProgrammeButtonModule from './EditDeliveryProgrammeButton';
 import type * as CreateDeliveryProgrammeButtonModule from './CreateDeliveryProgrammeButton';
+import { SnapshotFriendlyStylesProvider } from '../../utils';
 
-const EditDeliveryProgrammeButton: jest.MockedFn<typeof EditDeliveryProgrammeButtonModule['EditDeliveryProgrammeButton']> = jest.fn();
-const CreateDeliveryProgrammeButton: jest.MockedFn<typeof CreateDeliveryProgrammeButtonModule['CreateDeliveryProgrammeButton']> = jest.fn();
+const EditDeliveryProgrammeButton: jest.MockedFn<
+  (typeof EditDeliveryProgrammeButtonModule)['EditDeliveryProgrammeButton']
+> = jest.fn();
+const CreateDeliveryProgrammeButton: jest.MockedFn<
+  (typeof CreateDeliveryProgrammeButtonModule)['CreateDeliveryProgrammeButton']
+> = jest.fn();
 
 beforeEach(() => {
   jest.spyOn(global.Math, 'random').mockReturnValue(0);
@@ -62,7 +67,9 @@ describe('DeliveryProgrammeViewPageComponent', () => {
               [deliveryProgrammeApiRef, mockDeliveryProgrammeApi],
             ]}
           >
-            <DeliveryProgrammeViewPageComponent />
+            <SnapshotFriendlyStylesProvider>
+              <DeliveryProgrammeViewPageComponent />
+            </SnapshotFriendlyStylesProvider>
           </TestApiProvider>,
           {
             mountedRoutes: {
