+++ app/plugins/adp/src/components/DeliveryProject/EditDeliveryProjectButton.test.tsx
@@ -1,34 +1,48 @@
-import type { AlertApi} from '@backstage/core-plugin-api';
+import type { AlertApi } from '@backstage/core-plugin-api';
 import { alertApiRef } from '@backstage/core-plugin-api';
 import React from 'react';
-import type { DeliveryProjectApi} from './api';
+import type { DeliveryProjectApi } from './api';
 import { deliveryProjectApiRef } from './api';
 import { render as testRender, waitFor } from '@testing-library/react';
 import { TestApiProvider } from '@backstage/test-utils';
-import type {
-  CreateDeliveryProjectButtonProps} from './CreateDeliveryProjectButton';
-import {
-  CreateDeliveryProjectButton
-} from './CreateDeliveryProjectButton';
+import type { EditDeliveryProjectButtonProps } from './EditDeliveryProjectButton';
+import { EditDeliveryProjectButton } from './EditDeliveryProjectButton';
 import userEvent from '@testing-library/user-event';
-import type {
-  DeliveryProjectFields} from './DeliveryProjectFormFields';
-import {
-  DeliveryProjectFormFields,
-  emptyForm,
-} from './DeliveryProjectFormFields';
+import type { DeliveryProjectFields } from './DeliveryProjectFormFields';
+import { DeliveryProjectFormFields } from './DeliveryProjectFormFields';
 import { act } from 'react-dom/test-utils';
-import type { ValidationError as IValidationError } from '@internal/plugin-adp-common';
-import { ValidationError } from '../../utils';
+import type {
+  DeliveryProject,
+  ValidationError as IValidationError,
+} from '@internal/plugin-adp-common';
+import { SnapshotFriendlyStylesProvider, ValidationError } from '../../utils';
 import type * as PluginPermissionReactModule from '@backstage/plugin-permission-react';
 import type * as DialogFormModule from '../../utils/DialogForm';
 
 const usePermission: jest.MockedFn<
   typeof PluginPermissionReactModule.usePermission
 > = jest.fn();
-const DialogForm: jest.MockedFn<
-  typeof DialogFormModule.DialogForm
-> = jest.fn();
+const DialogForm: jest.MockedFn<typeof DialogFormModule.DialogForm> = jest.fn();
+
+const deliveryProject: DeliveryProject = {
+  ado_project: 'my ado project',
+  created_at: new Date(0),
+  delivery_programme_code: 'ADP',
+  delivery_programme_id: '00000000-0000-0000-0000-000000000001',
+  delivery_project_code: 'TRD',
+  description: 'My project',
+  id: '00000000-0000-0000-0000-000000000002',
+  name: 'my-cool-project',
+  namespace: 'ADP-TRD',
+  service_owner: 'someone else',
+  team_type: 'delivery',
+  title: 'My cool Project',
+  updated_at: new Date(0),
+  alias: 'ABC',
+  finance_code: '123',
+  github_team_visibility: 'public',
+  updated_by: 'Me',
+};
 
 const fields: DeliveryProjectFields = {
   ado_project: 'abc',
@@ -48,12 +62,15 @@ beforeEach(() => {
   jest.resetAllMocks();
 });
 
-describe('CreateDeliveryProjectButton', () => {
-  it('Should not render when the user is not allowed to create delivery projects', async () => {
+describe('EditDeliveryProjectButton', () => {
+  it('Should not render when the user is not allowed to edit delivery projects', async () => {
     const { mockAlertApi, mockProjectApi, render } = setup();
     usePermission.mockReturnValue({ allowed: false, loading: false });
 
-    const { result } = await render({ content: 'My button' });
+    const { result } = await render({
+      content: 'My button',
+      deliveryProject,
+    });
 
     expect(result.baseElement).toMatchSnapshot();
     expect(mockAlertApi.alert$).not.toHaveBeenCalled();
@@ -67,7 +84,10 @@ describe('CreateDeliveryProjectButton', () => {
     const { mockAlertApi, mockProjectApi, render } = setup();
     usePermission.mockReturnValue({ allowed: true, loading: false });
 
-    const { result } = await render({ content: 'My button' });
+    const { result } = await render({
+      content: 'My button',
+      deliveryProject,
+    });
 
     expect(result.baseElement).toMatchSnapshot();
     expect(mockAlertApi.alert$).not.toHaveBeenCalled();
@@ -82,8 +102,11 @@ describe('CreateDeliveryProjectButton', () => {
     usePermission.mockReturnValue({ allowed: true, loading: false });
     DialogForm.mockReturnValue(<span>This is a dialog!</span>);
 
-    const { result } = await render({ content: 'My button' });
-    await userEvent.click(result.getByTestId('create-delivery-project-button'));
+    const { result } = await render({
+      content: 'My button',
+      deliveryProject,
+    });
+    await userEvent.click(result.getByTestId('edit-delivery-project-button'));
 
     expect(result.baseElement).toMatchSnapshot();
     expect(DialogForm.mock.calls).toHaveLength(1);
@@ -99,12 +122,20 @@ describe('CreateDeliveryProjectButton', () => {
     }).toMatchObject({
       open: undefined,
       renderFields: DeliveryProjectFormFields,
-      confirm: 'Create',
+      confirm: 'Update',
       cancel: undefined,
       defaultValues: {
-        ...emptyForm,
-        github_team_visibility: 'public',
-        team_type: 'delivery',
+        ado_project: deliveryProject.ado_project,
+        delivery_programme_id: deliveryProject.delivery_programme_id,
+        delivery_project_code: deliveryProject.delivery_project_code,
+        description: deliveryProject.description,
+        github_team_visibility: deliveryProject.github_team_visibility,
+        namespace: deliveryProject.namespace,
+        service_owner: deliveryProject.service_owner,
+        team_type: deliveryProject.team_type,
+        title: deliveryProject.title,
+        alias: deliveryProject.alias,
+        finance_code: deliveryProject.finance_code,
       },
       disabled: {
         namespace: true,
@@ -126,8 +157,11 @@ describe('CreateDeliveryProjectButton', () => {
     usePermission.mockReturnValue({ allowed: true, loading: false });
     DialogForm.mockReturnValue(<span>This is a dialog!</span>);
 
-    const { result } = await render({ content: 'My button' });
-    await userEvent.click(result.getByTestId('create-delivery-project-button'));
+    const { result } = await render({
+      content: 'My button',
+      deliveryProject,
+    });
+    await userEvent.click(result.getByTestId('edit-delivery-project-button'));
 
     expect(result.baseElement).toMatchSnapshot('Before cancel');
     expect(DialogForm.mock.calls).toHaveLength(1);
@@ -144,20 +178,24 @@ describe('CreateDeliveryProjectButton', () => {
     expect(mockProjectApi.getDeliveryProjects).not.toHaveBeenCalled();
     expect(mockProjectApi.updateDeliveryProject).not.toHaveBeenCalled();
   });
-  it('Should call onCreated when the form closes with a value.', async () => {
+  it('Should call onEditd when the form closes with a value.', async () => {
     const { mockAlertApi, mockProjectApi, render } = setup();
     usePermission.mockReturnValue({ allowed: true, loading: false });
     DialogForm.mockReturnValue(<span>This is a dialog!</span>);
-    const onCreated: jest.MockedFn<
-      Exclude<CreateDeliveryProjectButtonProps['onCreated'], undefined>
+    const onEdited: jest.MockedFn<
+      Exclude<EditDeliveryProjectButtonProps['onEdited'], undefined>
     > = jest.fn();
 
-    const { result } = await render({ content: 'My button', onCreated });
-    await userEvent.click(result.getByTestId('create-delivery-project-button'));
+    const { result } = await render({
+      content: 'My button',
+      deliveryProject,
+      onEdited,
+    });
+    await userEvent.click(result.getByTestId('edit-delivery-project-button'));
 
     expect(result.baseElement).toMatchSnapshot('Before complete');
     expect(DialogForm.mock.calls).toHaveLength(1);
-    expect(onCreated).not.toHaveBeenCalled();
+    expect(onEdited).not.toHaveBeenCalled();
     const formProps = DialogForm.mock.calls[0][0];
     act(() => formProps.completed(fields));
     await waitFor(() =>
@@ -166,7 +204,7 @@ describe('CreateDeliveryProjectButton', () => {
     expect(result.baseElement).toMatchSnapshot('After complete');
     expect(mockAlertApi.alert$).not.toHaveBeenCalled();
     expect(mockAlertApi.post).not.toHaveBeenCalled();
-    expect(onCreated).toHaveBeenCalledTimes(1);
+    expect(onEdited).toHaveBeenCalledTimes(1);
     expect(mockProjectApi.createDeliveryProject).not.toHaveBeenCalled();
     expect(mockProjectApi.getDeliveryProjectById).not.toHaveBeenCalled();
     expect(mockProjectApi.getDeliveryProjects).not.toHaveBeenCalled();
@@ -177,8 +215,11 @@ describe('CreateDeliveryProjectButton', () => {
     usePermission.mockReturnValue({ allowed: true, loading: false });
     DialogForm.mockReturnValue(<span>This is a dialog!</span>);
 
-    const { result } = await render({ content: 'My button' });
-    await userEvent.click(result.getByTestId('create-delivery-project-button'));
+    const { result } = await render({
+      content: 'My button',
+      deliveryProject,
+    });
+    await userEvent.click(result.getByTestId('edit-delivery-project-button'));
 
     expect(result.baseElement).toMatchSnapshot();
     expect(DialogForm.mock.calls).toHaveLength(1);
@@ -192,17 +233,17 @@ describe('CreateDeliveryProjectButton', () => {
 
     const submitResult = await formProps.submit(fields);
     expect(submitResult).toMatchObject({ type: 'success' });
-    expect(mockProjectApi.createDeliveryProject.mock.calls).toMatchObject([
+    expect(mockProjectApi.updateDeliveryProject.mock.calls).toMatchObject([
       [fields],
     ]);
+    expect(mockProjectApi.createDeliveryProject).not.toHaveBeenCalled();
     expect(mockProjectApi.getDeliveryProjectById).not.toHaveBeenCalled();
     expect(mockProjectApi.getDeliveryProjects).not.toHaveBeenCalled();
-    expect(mockProjectApi.updateDeliveryProject).not.toHaveBeenCalled();
     expect(mockAlertApi.alert$).not.toHaveBeenCalled();
     expect(mockAlertApi.post.mock.calls).toMatchObject([
       [
         {
-          message: 'Delivery Project created successfully.',
+          message: 'Delivery Project updated successfully.',
           severity: 'success',
           display: 'transient',
         },
@@ -221,12 +262,15 @@ describe('CreateDeliveryProjectButton', () => {
         },
       },
     ];
-    mockProjectApi.createDeliveryProject.mockRejectedValueOnce(
+    mockProjectApi.updateDeliveryProject.mockRejectedValueOnce(
       new ValidationError(validationErrors),
     );
 
-    const { result } = await render({ content: 'My button' });
-    await userEvent.click(result.getByTestId('create-delivery-project-button'));
+    const { result } = await render({
+      content: 'My button',
+      deliveryProject,
+    });
+    await userEvent.click(result.getByTestId('edit-delivery-project-button'));
 
     expect(result.baseElement).toMatchSnapshot();
     expect(DialogForm.mock.calls).toHaveLength(1);
@@ -241,12 +285,12 @@ describe('CreateDeliveryProjectButton', () => {
       type: 'validationError',
       errors: validationErrors,
     });
-    expect(mockProjectApi.createDeliveryProject.mock.calls).toMatchObject([
+    expect(mockProjectApi.updateDeliveryProject.mock.calls).toMatchObject([
       [fields],
     ]);
+    expect(mockProjectApi.createDeliveryProject).not.toHaveBeenCalled();
     expect(mockProjectApi.getDeliveryProjectById).not.toHaveBeenCalled();
     expect(mockProjectApi.getDeliveryProjects).not.toHaveBeenCalled();
-    expect(mockProjectApi.updateDeliveryProject).not.toHaveBeenCalled();
     expect(mockAlertApi.alert$).not.toHaveBeenCalled();
     expect(mockAlertApi.post).not.toHaveBeenCalled();
   });
@@ -267,7 +311,7 @@ function setup() {
   return {
     mockAlertApi,
     mockProjectApi,
-    async render(props: CreateDeliveryProjectButtonProps) {
+    async render(props: EditDeliveryProjectButtonProps) {
       const result = testRender(
         <TestApiProvider
           apis={[
@@ -275,7 +319,9 @@ function setup() {
             [deliveryProjectApiRef, mockProjectApi],
           ]}
         >
-          <CreateDeliveryProjectButton {...props} />
+          <SnapshotFriendlyStylesProvider>
+            <EditDeliveryProjectButton {...props} />
+          </SnapshotFriendlyStylesProvider>
         </TestApiProvider>,
       );
       await waitFor(() => expect(result.baseElement).not.toBeEmptyDOMElement());
