+++ app/plugins/adp/src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useAsyncDataSource';
-export * from './useComputedUntilChanged';
 export * from './useArmsLengthBodyList';
 export * from './useDeliveryProgrammesList';
 export * from './useEntityRoute';
