+++ app/plugins/adp/src/utils/FormSelectField.tsx
@@ -6,12 +6,12 @@ import type {
   FieldPath,
   FieldPathValue,
   FieldValues,
-  UseControllerProps} from 'react-hook-form';
-import {
-  Controller
+  UseControllerProps,
 } from 'react-hook-form';
+import { Controller } from 'react-hook-form';
 import { isFieldDisabled } from './isFieldDisabled';
 import { enrichHelperText } from './enrichHelperText';
+import { rulesToHtmlProperties } from './rulesToHtmlProperties';
 
 export type FormSelectFieldProps<
   TFields extends FieldValues,
@@ -67,6 +67,9 @@ export function FormSelectField<
           }
           data-testid={name}
           disabled={isFieldDisabled(disabled, name)}
+          inputProps={{
+            ...rulesToHtmlProperties(rules),
+          }}
         >
           {options.map(x => (
             <MenuItem
