+++ app/plugins/adp/src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './TitleWithHelp';
 export * from './SelectedChipsRenderer';
 export * from './ValidationError';
 export * from './readValidationError';
+export * from './SnapshotFriendlyStylesProvider';
